Elements abstraced as Behaviour and surface types
