--- Intern/rayx-core/src/Beamline/Objects/Cone.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-struct DesignElement;  
-
-Element makeCone(const DesignElement& dele);
-
-}  // namespace RAYX
