--- Intern/rayx-core/src/Beamline/Objects/Cylinder.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-
-enum class CylinderDirection { LongRadiusR, ShortRadiusRho };
-struct DesignElement;  
-
-Element makeCylinder(const DesignElement& dele);
-
-}  // namespace RAYX
