--- Intern/rayx-core/src/Beamline/Objects/Ellipsoid.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-
-enum class FigureRotation { Yes, Plane, A11 };
-struct DesignElement;
-Element makeEllipsoid(const DesignElement& dele);
-
-}  // namespace RAYX
