--- Intern/rayx-core/src/Beamline/Objects/ImagePlane.cpp
@@ -1,14 +1,0 @@
-#include "ImagePlane.h"
-
-#include "Shader/Constants.h"
-
-namespace RAYX {
-
-Element makeImagePlane(const DesignElement& dele) {
-    auto behaviour = serializeImagePlane();
-    auto surface = serializePlaneXZ();
-    auto cutout = serializeUnlimited();
-    return makeElement(dele, behaviour, surface, cutout, DesignPlane::XY);
-}
-
-}  // namespace RAYX
