--- Intern/rayx-core/src/Beamline/Objects/ImagePlane.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-#include "DesignElement/DesignElement.h"
-
-namespace RAYX {
-struct DesignElement;  
-Element makeImagePlane(const DesignElement& dele);
-
-}  // namespace RAYX
