+++ Intern/rayx-core/src/Beamline/Objects/Objects.h
@@ -13,17 +13,3 @@ namespace RAYX{
 #include "Beamline/Objects/PixelSource.h"
 #include "Beamline/Objects/PointSource.h"
 #include "Beamline/Objects/SimpleUndulatorSource.h"
-
-// Optical Elements
-#include "Beamline/Objects/Cone.h"
-#include "Beamline/Objects/Cylinder.h"
-#include "Beamline/Objects/Ellipsoid.h"
-#include "Beamline/Objects/ImagePlane.h"
-#include "Beamline/Objects/Paraboloid.h"
-#include "Beamline/Objects/PlaneGrating.h"
-#include "Beamline/Objects/PlaneMirror.h"
-#include "Beamline/Objects/ReflectionZonePlate.h"
-#include "Beamline/Objects/Slit.h"
-#include "Beamline/Objects/SphereGrating.h"
-#include "Beamline/Objects/SphereMirror.h"
-#include "Beamline/Objects/ToroidMirror.h"
