--- Intern/rayx-core/src/Beamline/Objects/Paraboloid.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-
-struct DesignElement;
-enum class ParaboloidDirection { LongRadiusR, ShortRadiusRho };
-Element makeParaboloid(const DesignElement& dele);
-
-}  // namespace RAYX
