--- Intern/rayx-core/src/Beamline/Objects/PlaneGrating.h
@@ -1,10 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-#include "DesignElement/DesignElement.h"
-
-namespace RAYX {
-
-struct DesignElement;
-Element makePlaneGrating(const DesignElement& dele);
-
-}  // namespace RAYX
