--- Intern/rayx-core/src/Beamline/Objects/PlaneMirror.cpp
@@ -1,14 +1,0 @@
-#include "PlaneMirror.h"
-
-#include "Data/xml.h"
-#include "Shader/Constants.h"
-
-namespace RAYX {
-
-Element makePlaneMirror(const DesignElement& dele) {
-    auto behaviour = serializeMirror();
-    auto surface = makePlane();
-    return makeElement(dele, behaviour, surface);
-}
-
-}  // namespace RAYX
