--- Intern/rayx-core/src/Beamline/Objects/PlaneMirror.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-struct DesignElement;  
-
-Element makePlaneMirror(const DesignElement& dele);
-
-}  // namespace RAYX
