--- Intern/rayx-core/src/Beamline/Objects/SphereGrating.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-struct DesignElement;  
-Element makeSphereGrating(const DesignElement& dele);
-
-}  // namespace RAYX
