--- Intern/rayx-core/src/Beamline/Objects/SphereMirror.h
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace RAYX {
-
-struct DesignElement;  
-Element makeSphereMirror(const DesignElement& dele);
-
-}  // namespace RAYX
