--- Intern/rayx-core/src/Beamline/Objects/ToroidMirror.h
@@ -1,8 +1,0 @@
-#pragma once
-#include "Beamline/OpticalElement.h"
-
-namespace RAYX {
-struct DesignElement;  
-Element makeToroidMirror(const DesignElement& dele);
-
-}  // namespace RAYX
