+++ Intern/rayx-core/src/Data/DesignSourceWriter.h
@@ -69,6 +69,7 @@ void setMatrixSource(xml::Parser parser, DesignSource* ds) {
     setDefaultOrientation(parser, ds);
 }
 
+
 void setDipoleSource(xml::Parser parser, DesignSource* ds) {
     setAllMandatory(parser, ds);
