+++ Intern/rayx-core/src/Data/Importer.cpp
@@ -14,11 +14,12 @@
 #include "DesignElementWriter.h"
 #include "DesignSourceWriter.h"
 
+
+
 void parseElement(xml::Parser parser, DesignElement* de) {
     const char* type = parser.type();
-    // TODO add functions for each Element
 
-    if (strcmp(type, "ImagePlane") == 0) {
+    if (strcmp(type, "ImagePlane") == 0) {      
         getImageplane(parser, de);
     } else if (strcmp(type, "Cone") == 0) {
         getCone(parser, de);
@@ -28,8 +29,6 @@ void parseElement(xml::Parser parser, DesignElement* de) {
         getEllipsoid(parser, de);
     } else if (strcmp(type, "Experts Optics") == 0) {
         getExpertsOptics(parser, de);
-    } else if (strcmp(type, "Experts Cubic") == 0) {
-        getExpertsCubic(parser, de);
     } else if (strcmp(type, "Paraboloid") == 0) {
         getParaboloid(parser, de);
     } else if (strcmp(type, "Plane Grating") == 0) {
@@ -52,6 +51,7 @@ void parseElement(xml::Parser parser, DesignElement* de) {
         RAYX_WARN << "could not classify beamline object with Name: " << parser.name() << "; Type: " << parser.type();
     }
 }
+
 
 namespace RAYX {
 
@@ -95,7 +95,6 @@ void addBeamlineObjectFromXML(rapidxml::xml_node<>* node, Beamline* beamline, co
     } else if (strcmp(type, "Simple Undulator") == 0) {
         setSimpleUndulatorSource(parser, &ds);
         beamline->m_DesignSources.push_back(ds);
-        // addLightSource(std::make_shared<SimpleUndulatorSource>(parser), node);
     } else {
         parseElement(parser, &de);
         beamline->m_DesignElements.push_back(de);
