+++ Intern/rayx-core/src/DesignElement/Value.h
@@ -28,7 +28,8 @@ enum class ValueType {
     EnergyDistributionType,
     EnergySpreadUnit,
     ElectronEnergyOrientation,
-    SigmaType
+    SigmaType,
+    BehaviourType
 };
 
 class Undefined {};
@@ -70,6 +71,7 @@ class Value {
     Value(EnergyDistributionType x) : m_variant(x) {}
     Value(EnergySpreadUnit x) : m_variant(x) {}
     Value(SigmaType x) : m_variant(x) {}
+    Value(BehaviourType x) : m_variant(x) {}
 
     void operator=(double x) { m_variant = x; }
     void operator=(int x) { m_variant = x; }
@@ -93,6 +95,7 @@ class Value {
     void operator=(EnergySpreadUnit x) { m_variant = x; }
     void operator=(ElectronEnergyOrientation x) { m_variant = x; }
     void operator=(SigmaType x) { m_variant = x; }
+    void operator=(BehaviourType x) { m_variant = x; }
 
 
 
@@ -103,7 +106,7 @@ class Value {
             ValueType::Rad,       ValueType::Material,   ValueType::Misalignment,  ValueType::CentralBeamStop,
             ValueType::Cutout,    ValueType::Bool,       ValueType::FigureRotation,ValueType::CurvatureType,
             ValueType::Surface,   ValueType::SourceDist, ValueType::SpreadType,    ValueType::EnergyDistributionType,
-            ValueType::Dmat4x4,   ValueType::SigmaType,  ValueType::ElectronEnergyOrientation           
+            ValueType::Dmat4x4,   ValueType::SigmaType,  ValueType::BehaviourType, ValueType::ElectronEnergyOrientation,           
         };
         return types[m_variant.index()];
     }
@@ -240,11 +243,26 @@ class Value {
         return *x;
     }
 
+    inline BehaviourType as_behaviourType() const {
+        auto* x = std::get_if<BehaviourType>(&m_variant);
+        if (!x) throw std::runtime_error("as_behaviourType() called on non-behaviourType!");
+        return *x;
+    }
+
     const Value& operator[](std::string s) const {
         const Map* m = std::get_if<Map>(&m_variant);
-        if (!m) throw std::runtime_error("Indexing into non-map!");
-
-        return *m->at(s).get();  // TODO return undefined on missing
+        if (!m) throw std::runtime_error("Indexing into non-map at: " + s);
+        try
+        {
+            auto x = *m->at(s).get();
+        }
+        catch(const std::exception& e)
+        {
+            std::cerr << e.what() << '\n';
+            RAYX_LOG << "Indexing into non-map at: " << s ;
+        }
+    
+        return *m->at(s).get();
     }
 
     Value& operator[](std::string s) {
@@ -256,17 +274,17 @@ class Value {
         if (!m->contains(s)){
             (*m)[s] = std::make_shared<Value>();
         }
-        return *((*m)[s].get());
+        return *((*m)[s].get()); 
     }
 
   private:
     std::variant<
-                 Undefined,        double,          int,     ElectronEnergyOrientation,
-                 glm::dvec4,       glm::dmat4x4,    bool,    EnergyDistributionType, 
-                 Misalignment,     CentralBeamstop, Cutout,  CylinderDirection, 
-                 FigureRotation,   Map,             Surface, CurvatureType,
-                 SourceDist,       SpreadType,      Rad,     Material,
-                 EnergySpreadUnit, std::string,     SigmaType
+                 Undefined,        double,          int,       ElectronEnergyOrientation,
+                 glm::dvec4,       glm::dmat4x4,    bool,      EnergyDistributionType, 
+                 Misalignment,     CentralBeamstop, Cutout,    CylinderDirection, 
+                 FigureRotation,   Map,             Surface,   CurvatureType,
+                 SourceDist,       SpreadType,      Rad,       Material,
+                 EnergySpreadUnit, std::string,     SigmaType, BehaviourType
                 > m_variant;
 };
 }  // namespace RAYX
