+++ Intern/rayx-ui/src/Application.h
@@ -21,6 +21,8 @@ const bool enableValidationLayers = true;
 class Window;
 class Device;
 class RenderObject;
+struct DesignElement;
+struct DesignSource;
 
 class Application {
   public:
