+++ Intern/rayx-ui/src/Vertex.h
@@ -6,6 +6,9 @@
 #include <glm/glm.hpp>
 #include <variant>
 
+#include "DesignElement/DesignElement.h"
+#include "DesignElement/DesignSource.h"
+
 /**
  * @struct Vertex
  * @brief A structure representing a Vertex in 3D space with additional color information.
